+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- fix: ignore additional keyboard windows for $screen event ([#279](https://github.com/PostHog/posthog-ios/pull/279))
+
 ## 3.17.1 - 2024-12-18
 
 - fix: avoid masking SwiftUI Gradient views ([#275](https://github.com/PostHog/posthog-ios/pull/275))